+++ index.html
@@ -242,7 +242,7 @@
     </section>
   </template>
 
-  <script src="js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
 </body>
 
 </html>
